+++ src/views/Index.vue
@@ -1,0 +1,10 @@
+<template>
+  <div>
+    <div class="row">
+      <div class="col">
+        <h3 class="mt-4">What is 143chars?</h3>
+        <p>p 143chars is totally not a twitter clone. It is a project to learn more about Node.js, Express, and MongoDB.</p>
+      </div>
+    </div>
+  </div>
+</template>