+++ changelog.md
@@ -3,3 +3,6 @@
 ## 5/5/19
 * Added Bootstrap via npm
 * Added registration via API
+* Created Index and Register views
+* Created Navbar component
+* Created routes for / and /register